Refactor: move upload models to data.models
