+++ app/src/main/java/fr/free/nrw/commons/data/models/upload/UploadResponse.kt
@@ -1,3 +1,3 @@
-package fr.free.nrw.commons.upload
+package fr.free.nrw.commons.data.models.upload
 
 class UploadResponse(val upload: UploadResult?)
