+++ app/src/main/java/fr/free/nrw/commons/db/Converters.java
@@ -8,7 +8,7 @@ import fr.free.nrw.commons.CommonsApplication;
 import fr.free.nrw.commons.contributions.ChunkInfo;
 import fr.free.nrw.commons.di.ApplicationlessInjection;
 import fr.free.nrw.commons.location.LatLng;
-import fr.free.nrw.commons.upload.WikidataPlace;
+import fr.free.nrw.commons.data.models.upload.WikidataPlace;
 import fr.free.nrw.commons.upload.structure.depictions.DepictedItem;
 import java.util.Date;
 import java.util.List;
