+++ app/src/main/java/fr/free/nrw/commons/description/DescriptionEditActivity.kt
@@ -15,7 +15,7 @@ import fr.free.nrw.commons.description.EditDescriptionConstants.UPDATED_WIKITEXT
 import fr.free.nrw.commons.description.EditDescriptionConstants.WIKITEXT
 import fr.free.nrw.commons.kvstore.JsonKvStore
 import fr.free.nrw.commons.settings.Prefs
-import fr.free.nrw.commons.upload.UploadMediaDetail
+import fr.free.nrw.commons.data.models.upload.UploadMediaDetail
 import fr.free.nrw.commons.upload.UploadMediaDetailAdapter
 import fr.free.nrw.commons.utils.DialogUtil.showAlertDialog
 import java.util.*
