+++ app/src/main/java/fr/free/nrw/commons/repository/UploadRepository.java
@@ -9,10 +9,10 @@ import fr.free.nrw.commons.filepicker.UploadableFile;
 import fr.free.nrw.commons.location.LatLng;
 import fr.free.nrw.commons.nearby.NearbyPlaces;
 import fr.free.nrw.commons.nearby.Place;
-import fr.free.nrw.commons.upload.ImageCoordinates;
+import fr.free.nrw.commons.data.models.upload.ImageCoordinates;
 import fr.free.nrw.commons.upload.SimilarImageInterface;
 import fr.free.nrw.commons.upload.UploadController;
-import fr.free.nrw.commons.upload.UploadItem;
+import fr.free.nrw.commons.data.models.upload.UploadItem;
 import fr.free.nrw.commons.upload.UploadModel;
 import fr.free.nrw.commons.upload.structure.depictions.DepictModel;
 import fr.free.nrw.commons.upload.structure.depictions.DepictedItem;
