+++ app/src/main/java/fr/free/nrw/commons/upload/FileUtils.java
@@ -7,6 +7,7 @@ import android.webkit.MimeTypeMap;
 
 import androidx.exifinterface.media.ExifInterface;
 
+import fr.free.nrw.commons.data.models.upload.ImageCoordinates;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
