+++ app/src/main/java/fr/free/nrw/commons/upload/SimilarImageInterface.java
@@ -1,5 +1,7 @@
 package fr.free.nrw.commons.upload;
 
+import fr.free.nrw.commons.data.models.upload.ImageCoordinates;
+
 public interface SimilarImageInterface {
     void showSimilarImageFragment(String originalFilePath, String possibleFilePath,
         ImageCoordinates similarImageCoordinates);
