+++ app/src/main/java/fr/free/nrw/commons/upload/UploadClient.java
@@ -9,6 +9,11 @@ import com.google.gson.Gson;
 import fr.free.nrw.commons.CommonsApplication;
 import fr.free.nrw.commons.contributions.ChunkInfo;
 import fr.free.nrw.commons.contributions.Contribution;
+import fr.free.nrw.commons.data.models.upload.CountingRequestBody;
+import fr.free.nrw.commons.data.models.upload.StashUploadResult;
+import fr.free.nrw.commons.data.models.upload.StashUploadState;
+import fr.free.nrw.commons.data.models.upload.UploadResponse;
+import fr.free.nrw.commons.data.models.upload.UploadResult;
 import fr.free.nrw.commons.upload.worker.UploadWorker.NotificationUpdateProgressListener;
 import io.reactivex.Observable;
 import io.reactivex.disposables.CompositeDisposable;
