+++ app/src/main/java/fr/free/nrw/commons/upload/UploadInterface.java
@@ -3,6 +3,7 @@ package fr.free.nrw.commons.upload;
 import androidx.annotation.NonNull;
 
 import com.google.gson.JsonObject;
+import fr.free.nrw.commons.data.models.upload.UploadResponse;
 import io.reactivex.Observable;
 import okhttp3.MultipartBody;
 import okhttp3.RequestBody;
