+++ app/src/main/java/fr/free/nrw/commons/upload/depicts/DepictsDao.kt
@@ -1,6 +1,7 @@
 package fr.free.nrw.commons.upload.depicts
 
 import androidx.room.*
+import fr.free.nrw.commons.data.models.upload.depictions.Depicts
 import fr.free.nrw.commons.upload.structure.depictions.DepictedItem
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
