+++ app/src/main/java/fr/free/nrw/commons/wikidata/WikiBaseClient.java
@@ -3,7 +3,7 @@ package fr.free.nrw.commons.wikidata;
 import static fr.free.nrw.commons.media.MediaClientKt.PAGE_ID_PREFIX;
 import static fr.free.nrw.commons.di.NetworkingModule.NAMED_COMMONS_CSRF;
 
-import fr.free.nrw.commons.upload.UploadResult;
+import fr.free.nrw.commons.data.models.upload.UploadResult;
 import fr.free.nrw.commons.upload.WikiBaseInterface;
 import io.reactivex.Observable;
 import javax.inject.Inject;
