+++ app/src/main/java/fr/free/nrw/commons/wikidata/WikidataEditService.java
@@ -9,14 +9,13 @@ import androidx.annotation.Nullable;
 import com.google.gson.Gson;
 import fr.free.nrw.commons.R;
 import fr.free.nrw.commons.contributions.Contribution;
+import fr.free.nrw.commons.data.models.upload.UploadResult;
 import fr.free.nrw.commons.kvstore.JsonKvStore;
-import fr.free.nrw.commons.upload.UploadResult;
-import fr.free.nrw.commons.upload.WikidataItem;
-import fr.free.nrw.commons.upload.WikidataPlace;
+import fr.free.nrw.commons.data.models.upload.WikidataItem;
+import fr.free.nrw.commons.data.models.upload.WikidataPlace;
 import fr.free.nrw.commons.utils.ConfigUtils;
 import fr.free.nrw.commons.utils.ViewUtil;
 import io.reactivex.Observable;
-import io.reactivex.disposables.Disposable;
 import io.reactivex.schedulers.Schedulers;
 import java.util.ArrayList;
 import java.util.Arrays;
