+++ app/src/test/kotlin/fr/free/nrw/commons/contributions/ContributionsListFragmentUnitTests.kt
@@ -16,7 +16,7 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton
 import fr.free.nrw.commons.Media
 import fr.free.nrw.commons.TestAppAdapter
 import fr.free.nrw.commons.TestCommonsApplication
-import fr.free.nrw.commons.upload.WikidataPlace
+import fr.free.nrw.commons.data.models.upload.WikidataPlace
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
