+++ app/src/test/kotlin/fr/free/nrw/commons/upload/ImageProcessingServiceTest.kt
@@ -1,6 +1,8 @@
 package fr.free.nrw.commons.upload
 
 import android.net.Uri
+import fr.free.nrw.commons.data.models.upload.UploadItem
+import fr.free.nrw.commons.data.models.upload.UploadMediaDetail
 import fr.free.nrw.commons.location.LatLng
 import fr.free.nrw.commons.media.MediaClient
 import fr.free.nrw.commons.nearby.Place
